requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
numpy==2.1.3
pillow==11.0.0
opencv-python-headless==4.10.0.84
torch>=2.5.1
huggingface_hub>=0.24.0
git+https://github.com/facebookresearch/sam2.git